requirements.txt
absl-py==1.0.0
affine==2.3.0
albumentations==1.1.0
analytics-python==1.4.0
apache-beam==2.34.0
appnope==0.1.2
arrow==1.2.1
astroid==2.9.0
astunparse==1.6.3
attrs==21.2.0
avro-python3==1.9.2.1
backcall==0.2.0
backoff==1.10.0
boto3==1.20.13
botocore==1.23.13
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
commonmark==0.9.1
configparser==5.1.0
crcmod==1.7
cycler==0.11.0
cytoolz==0.11.2
dataclasses-json==0.5.6
debugpy==1.5.1
decorator==5.1.0
dill==0.3.1.1
docker-pycreds==0.4.0
docopt==0.6.2
efficientnet==1.0.0
entrypoints==0.3
fastavro==1.4.7
flatbuffers==2.0
fonttools==4.28.2
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.2.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
gql==3.0.0b1
graphql-core==3.1.6
grpcio==1.42.0
h5py==3.6.0
hdfs==2.6.0
httplib2==0.19.1
humanize==3.12.0
idna==3.3
image-classifiers==1.0.0
imageio==2.12.0
importlib-metadata==4.8.2
iniconfig==1.1.1
ipykernel==6.5.1
ipython==7.29.0
isort==5.10.1
jedi==0.18.1
jmespath==0.10.0
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
libclang==12.0.0
lightning-grid==0.5.1
Markdown==3.3.6
marshmallow==3.14.1
marshmallow-enum==1.5.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mccabe==0.6.1
memory-profiler==0.58.0
monotonic==1.6
multidict==5.2.0
mypy-extensions==0.4.3
nest-asyncio==1.5.1
networkx==2.6.3
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
orjson==3.6.4
packaging==21.3
pandas==1.3.4
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.11
platformdirs==2.4.0
pluggy==1.0.0
pretty-errors==1.2.25
promise==2.3
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
Pygments==2.10.0
PyJWT==1.7.1
pylint==2.12.1
pymongo==3.12.1
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-snappy==0.6.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==5.4.1
pyzmq==22.3.0
qudida==0.0.4
rasterio==1.2.10
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rich==10.14.0
rsa==4.8
s3transfer==0.5.0
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.3
segmentation-models==1.0.1
sentry-sdk==1.5.0
setuptools-scm==6.3.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
snuggs==1.4.7
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typeguard==2.13.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
tzdata==2021.5
tzlocal==4.1
ulid-py==1.1.0
urllib3==1.26.7
wandb==0.12.1
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.0.2
wrapt==1.13.3
yarg==0.1.9
yarl==1.7.2
yaspin==0.18.0
zipp==3.6.0
